Formula::parse
